==> notebooks/02.Procress semistructured data (json) to structured data (csv).ipynb <==
# Import data from json file
- input: jason files
- output: 
    - data1: list of json objects
    - data2: normailized dataframe of data
    - data3: 1000 sample records, with only relevant columns

# IMPORT JSON FILES TO A LIST OF OBJECTS
import json
with open('../data/1.pulledTweets.json','r') as f:
    data1=[json.loads(line) for line in f]

# CONVERT AND NORMALIZE DATA TO A PANDAS DATAFRAME
import pandas
from pandas.io.json import json_normalize
data2=pandas.json_normalize(data1)
data2.head(3)

# KEEP ONLY INTERESTED COLUMNS, write to csv file
data3=data2[['created_at','id','text']]
data3.to_csv('../data/2.pulledTweets.csv')
data3.tail(3)
