==> notebooks/03.Detect languages of documents and clear non-english documents.ipynb <==
# Language detection and clear non-english articles
- input: 
    - data3: dataframe with 'text' columns, in many languages
- output: 
    - data3: dataframe with only english articles

import pandas
data3=pandas.read_csv('../data/2.pulledTweets.csv')
data3.tail()

# GOOGLE LANGUAGE DETECTION

data3['language']=''
from langdetect import detect
for i in range(len(data3)):
    try: data3['language'][i]=detect(data3['text'][i])
    except: 
        data3['language'][i]='en'
        continue
data3.tail()

# Check out the number of messages in each language, top 5
data3[['language','id']].groupby('language').count().sort_values(by=['id'], ascending=False).head()

# DROP NON-ENGLISH ARTICLES
data3=data3[(data3['language']=='en')|(data3['language']=='')].reset_index()
data3[['language','id']].groupby('language').count()

data3.tail()

data3=data3[['created_at','id','text','language']]
data3.tail()

data3.to_csv('../data/3.pulledTweets-English.csv')
