==> notebooks/04.Clear spam messages with a supervised machine learning model.ipynb <==
## Build supervised spam detector

- Dataset to be used for traning: "Spam Training.csv"
- Fit and transform the training data X_train using a Count Vectorizer ignoring terms that have a document frequency strictly lower than **5** and using **character n-grams from n=2 to n=5.**

- To tell Count Vectorizer to use character n-grams pass in `analyzer='char_wb'` which creates character n-grams only from text inside word boundaries. This should make the model more robust to spelling mistakes.

- Using this document-term matrix and the following additional features:
    * the length of document (number of characters)
    * number of digits per document
    * **number of non-word characters (anything other than a letter, digit or underscore.)**

- fit a Logistic Regression model with regularization C=100. Then compute the area under the curve (AUC) score using the transformed test data.

#### Spam Model

def spam_model():
    import pandas as pd
    from sklearn.feature_extraction.text import CountVectorizer
    from sklearn.linear_model import LogisticRegression
    from sklearn.model_selection import train_test_split
    from sklearn.metrics import roc_auc_score
    import re
    import numpy as np
    from scipy.sparse import csr_matrix, hstack

    # Import and tranform spam dataset
    spam_data = pd.read_csv('../data/4.Spam Training.csv')
    spam_data['target'] = np.where(spam_data['target']=='spam',1,0)
    X_train, X_test, y_train, y_test = train_test_split(spam_data['text'], spam_data['target'], random_state=0)

    # vectorize train, test sets
    spam_vect=CountVectorizer(min_df=5, analyzer='char_wb', ngram_range=(2,5)).fit(X_train)
    X_train_vectorized = spam_vect.transform(X_train)
    X_test_vectorized = spam_vect.transform(X_test)

    # add features
    num_digit_X_train=X_train.apply(lambda x:len(re.findall(r'\d', x)))
    num_non_word_X_train=X_train.apply(lambda x:len(re.findall(r'\W', x)))
    X_train_vectorized_len_digit_nonword=hstack([X_train_vectorized, csr_matrix([X_train.str.len(), 
                                                                                 num_digit_X_train, 
                                                                                 num_non_word_X_train]).T], 'csr')

    num_digit_X_test=X_test.apply(lambda x:len(re.findall(r'\d', x)))
    num_non_word_X_test=X_test.apply(lambda x:len(re.findall(r'\W', x)))
    X_test_vectorized_len_digit_nonword=hstack([X_test_vectorized,csr_matrix([X_test.str.len(),
                                                                            num_digit_X_test,
                                                                            num_non_word_X_test]).T],'csr')

    # Creat prediction model
    spam_model=LogisticRegression(C=100).fit(X_train_vectorized_len_digit_nonword,y_train)

    # Evaluate the model
    feature_names = np.array(spam_vect.get_feature_names()+['length_of_doc', 'digit_count', 'non_word_char_count'])
    sorted_coef_index = spam_model.coef_[0].argsort()
    smallest = feature_names[sorted_coef_index[:10]]
    largest = feature_names[sorted_coef_index[:-11:-1]]
    spam_predictions = spam_model.predict(X_test_vectorized_len_digit_nonword)
    print('test score:', roc_auc_score(y_test, spam_predictions))
    print('list of smallest coef index:', list(smallest))
    print('list of largest coef index:', list(largest))
    
    return spam_vect, spam_model

spam_vect, spam_model=spam_model()

def spam_predict(docs):
    from scipy.sparse import csr_matrix, hstack
    import re
    docs_vectorized = spam_vect.transform(docs)
    num_digit=docs.apply(lambda x:len(re.findall(r'\d', x)))
    num_nonword=docs.apply(lambda x:len(re.findall(r'\W', x)))
    docs_vectorized_len_digit_nonword=hstack([docs_vectorized,csr_matrix([docs.str.len(),
                                                                    num_digit,
                                                                    num_nonword]).T],'csr')
    return spam_model.predict(docs_vectorized_len_digit_nonword)

###### Note: Need better spam dataset, or try different spam detection model later

import pandas as pd
df=pd.read_csv('../data/3.pulledTweets-English.csv').drop_duplicates(subset=['text']).reset_index()
df.text

df=df[df.text!='']
df

def clean_spams(file):
    import pandas as pd
    df=pd.read_csv(file).drop_duplicates(subset=['text']).reset_index()
   
    import re
    for i in range(len(df)):
        try:
            txt = df.loc[i]["text"]
            txt=re.sub(r'@[A-Z0-9a-z_:]+','',txt)#replace username-tags
            txt=re.sub(r'^[RT]+','',txt)#replace RT-tags
            txt = re.sub('https?://[A-Za-z0-9./]+','',txt)#replace URLs
            txt=re.sub("[^a-zA-Z]", " ",txt)#replace hashtags
            df.at[i,"text"]=txt
        except:
            df.at[i,"text"]=''
            continue

    df['spam']=spam_predict(df['text'])
    df.to_csv('../data/4.Tweet data before clear spams.csv')
    del df['Unnamed: 0']
    df=df[df['spam']==0]
    del df['spam']
    df=df.where(df.text!='')
    df.to_csv('../data/4.Tweet data after clear spams.csv')
    return df['text']

clean_spams('../data/3.pulledTweets-English.csv')
