==> notebooks/05.2.Remove similar documents using Locality Sensitive Hashing (LSH).ipynb <==
*This part is overkilled for tweets short-texts. At the end you will see non of the messages removed, but I want to put it here for a comprehensive text-mining demonstration. This part is more appropriate for longer documents like News*

# Deduplication using LSH
- input: 
    - data3: dataframe with 'text' english columns
- output: 
    - data4: dataframe without similar articles, based on 'text' columns

import pandas
data3=pandas.read_csv('../data/4.Tweet data after clear spams.csv')
data3.tail()

import os
import re
import random
import time
import binascii
from bisect import bisect_right
from heapq import heappop, heappush
import sys
from __future__ import division

# Convert documents to sets of shingles (shingle=3 consecutive words in documents)

data3['shingleSet']=''
totalShingles = 0
numDocs=len(data3)
for i in range(0, numDocs):
    words = str(data3.text[i]).split(" ") 
    shingleSet = set()
    for index in range(0, len(words) - 2):
        shingle = words[index] + " " + words[index + 1] + " " + words[index + 2]
        crc = binascii.crc32(shingle.encode()) & 0xffffffff # Hash the shingle to a 32-bit integer
        shingleSet.add(crc)
    data3['shingleSet'][i] = shingleSet
    totalShingles = totalShingles + len(shingleSet)

data3.tail()

totalShingles /numDocs #Average shingles per doc

# When comparing 2 documents, instead of compairing all shingles, we only compair random 30 shingles, by converting each document to a signature of length 30
numHashes = 10 #The length of signature designed for each document, make a justment based on average shingles per doc

numElems = int(numDocs * (numDocs - 1) / 2) # number of paired documents to compare similarities
estJSim = [0 for x in range(numElems)] #'estJSim' will be for the estimated Jaccard Similarities found by comparing the MinHash signatures.

# Define a function to map a 2D matrix coordinate into a 1D index.
def getTriangleIndex(i, j):
    if i == j:
        sys.stderr.write("Can't access triangle matrix with i == j")
        sys.exit(1)
    if j < i:
        temp = i
        i = j
        j = temp
    return int(i * (numDocs - (i + 1) / 2.0) + j - i) - 1

maxShingleID = 2**32-1
def pickRandomCoeffs(k):  # Create a list of 'k' random values between 0 and maxShingleID
    randList = []
    while k > 0:
        randIndex = random.randint(0, maxShingleID) 
        while randIndex in randList:
            randIndex = random.randint(0, maxShingleID) 
        randList.append(randIndex)
        k = k - 1
    return randList

# Generating random hash functions

nextPrime = 4294967311 # The prime number above 'maxShingleID'

coeffA,coeffB = pickRandomCoeffs(numHashes),pickRandomCoeffs(numHashes) #each pair corresponse to a function that creat the first element in each signature

# Generating MinHash signatures for all documents

data3['signature'] = ''

for index in range(0,numDocs):
    shingleSet = data3['shingleSet'][index]
    signature = []
    for i in range(0, numHashes): #creating each elemment of a signature
        minHashCode = nextPrime + 1
        for shingle in shingleSet:
            hashCode = (coeffA[i] * shingle + coeffB[i]) % nextPrime 
            if hashCode < minHashCode:# Track the lowest hash code seen
                minHashCode = hashCode
        signature.append(minHashCode)
    data3['signature'][index] = signature

data3.head()

# Generate pairs of duplicate articles
threshold = 0.5  

docA = []  #first document in the duplicate pair - store uuid here
docB = []  #second document in the duplicate pair - store uuid here
for i in range(0, numDocs):  
    for j in range(i + 1, numDocs):
        estJ = estJSim[getTriangleIndex(i, j)]
        if estJ > threshold:
            docA.append(i)
            docB.append(j)
    
    #   Calculate the actual Jaccard similarity for validation. Comment out this part when you done checking
            # s1 = data3['shingleSet'][i]
            # s2 = data3['shingleSet'][j]
            # if len(s1.union(s2)) != 0:
            #     J = (len(s1.intersection(s2)) / len(s1.union(s2)))
            # print ("  %5s --> %5s   %.2f     %.2f" % (i, j, estJ, J))

# data3.loc[docA].text.to_csv('docA.csv')
# data3.loc[docB].text.to_csv('docB.csv')

# Remove duplicate documents
docs_to_remove = set(docB)
data4=data3.drop(docB)
data4.to_csv('../data/5.pulledTweet-deduplicated.csv')
data4.tail()
