==> notebooks/06.Extract word frequency & summarize text.ipynb <==
# Word Frequency
- input: 
    - data4: dataframe with 'text' columns
- output: 
    - data4: added 'text_clearn' columns which is spacy tokens of 'text' columns

import pandas
data4=pandas.read_csv('../data/5.pulledTweet-deduplicated.csv')
data4.tail()

# Generate clean tokens for all documents

import spacy
nlp = spacy.load('en_core_web_sm', disable=['parser', 'ner']) #just keep tagger for lemmatization, disable 'parser' and 'ner'

def token_filter(token):
    return not (token.is_punct | token.is_space | token.is_stop | len(token.text) <= 2)

def spacy_tokenizer(text):
    for doc in nlp.pipe([str(text).lower()]):
        tokens = [token.lemma_ for token in doc if token_filter(token)]
    return tokens

data4['token'] = data4['text'].map(spacy_tokenizer) 
data4['text_clean'] = data4.apply(lambda x: " ".join(x['token']), axis=1)

data4.tail()

import nltk
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

def get_top_phrase(corpus, numPhrases=None, numWords=1):
    
    vec = CountVectorizer(ngram_range=(numWords, numWords), stop_words=nltk.corpus.stopwords.words('english')).fit(corpus.astype(str))
    bag_of_words = vec.transform(corpus)
    sum_words = bag_of_words.sum(axis=0) 
    words_freq = [(word, sum_words[0, idx]) for word, idx in vec.vocabulary_.items()]
    words_freq =sorted(words_freq, key = lambda x: x[1], reverse=True)
    return words_freq[:numPhrases]

common_words = get_top_phrase(data4['text_clean'], numPhrases=10,numWords=4)
pandas.DataFrame(common_words, columns = ['bigram' , 'count'])

# Text summarization using word frequency (extractive approach)

# Join all articles into a big string
big_doc=''
for i in range(len(data4)):
    big_doc+=str(data4.text[i])+'.'

# OPTION 1 #very slow for alot of big documents
from gensim.summarization import summarize 
%time
summarize(big_doc, ratio = 0.02)

# OPTION 2
import nltk
from nltk.tokenize import word_tokenize, sent_tokenize

%time

sentences = sent_tokenize(big_doc)
sentences = [sentence for sentence in sentences if len(sentence) >= 30]
sentences = list(set(sentences))
freq_table = dict(get_top_phrase(data4['text'].astype(str),numWords=1))

def score_sentences(sentences, freq_table):
    sentence_scores = {}
    for sentence in sentences:
        word_count_in_sentence = len(word_tokenize(sentence))
        for word in freq_table:
            if word.lower() in sentence.lower():                
                if sentence in sentence_scores:
                    sentence_scores[sentence] += freq_table[word]
                else:
                    sentence_scores[sentence] = freq_table[word]
        sentence_scores[sentence] = sentence_scores[sentence] // word_count_in_sentence
    return sentence_scores
sentence_scores = score_sentences(sentences, freq_table)


def average_sentence_scores(sentence_scores):
    sum=0
    for key, value in sentence_scores.items():
        sum+=value
    return sum/len(sentence_scores)
print('Average sentence scores is: ',average_sentence_scores(sentence_scores),'. Triple it to use as threahold for summary funciton below?')

def generate_summary(sentences, sentence_scores, threshold): # Use average_sentence_scores as clue for threshold
    sentence_count = 0
    summary = ''
    for sentence in sentences:
        if sentence in sentence_scores and sentence_scores[sentence] > threshold:
            summary += " " + sentence
            sentence_count += 1
    return summary 

generate_summary(sentences, sentence_scores, 6)
