==> notebooks/07.Recognize entities .ipynb <==
# Entity Recognition

import pandas
data4=pandas.read_csv('../data/5.pulledTweet-deduplicated.csv')
data4.tail()

import spacy
nlp = spacy.load("en_core_web_sm", disable = ['tagger','lemmatizer','parser'])
ents = []
for doc in data4['text']:
    doc=str(doc)
    doc = nlp(doc)
    for ent in doc.ents:
            if ent.label_ in ['ORG','PERSON','FAC','GPE']:
                ents.append(str(ent))

from collections import Counter
Counter(ents).most_common(10)
