==> notebooks/08.Build topics model from text.ipynb <==
*This part is more appropriate to use for long documents like News. I put this here for a comprehensive text-mining demonstration only*

# Topic Modeling using Gensism
Build topics model from the given messages in 'data/5.pulledTweet-deduplicated.csv'

import pandas
data4=pandas.read_csv('../data/5.pulledTweet-deduplicated.csv')
data4.tail()

import gensim
from sklearn.feature_extraction.text import CountVectorizer

# Use CountVectorizor to find three letter tokens, remove stop_words, 
# remove tokens that don't appear in at least x documents,
# remove tokens that appear in more than y% of the documents
vect = CountVectorizer(min_df=1, max_df=0.4, stop_words='english')

X = vect.fit_transform(data4['text'].astype(str))

# Convert sparse matrix to gensim corpus.
corpus = gensim.matutils.Sparse2Corpus(X, documents_columns=False)

# Mapping from word IDs to words (To be used in LdaModel's id2word parameter)
id_map = dict((v, k) for k, v in vect.vocabulary_.items())

# Use the gensim.models.ldamodel.LdaModel constructor to estimate 
# LDA model parameters on the corpus, and save to the variable `ldamodel`

ldamodel = gensim.models.ldamodel.LdaModel (corpus, id2word=id_map, passes=5)

#### Print Most important topics, select number of topics and number of words in each topics you want

ldamodel.print_topics(num_topics=5, num_words=5)

#### Topic distriution
Given a document, this function returns a list of tuples, where each tuple is `(#topic, probability)`*

def topic_distribution(doc:str):
    doc=[doc]
    docVectorized = vect.transform(doc)
    docCorpus = gensim.matutils.Sparse2Corpus(docVectorized, documents_columns=False)
    result=[]
    for item in ldamodel[docCorpus]:
        result+=item
    return result

topic_distribution(" We understand your concern  Some items may get shipped separately  To clarify  have we missed the delivery  ")

ldamodel.print_topic(topicno=34,topn=7)
